Refactor the code to follow software practices.
