+++ package.json
@@ -1,0 +1,16 @@
+{
+  "name": "coding-challenge-backend",
+  "version": "1.0.0",
+  "description": "Creating a reat estate purchasing app",
+  "main": "index.js",
+  "scripts": {
+    "start": "node App.js",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "readline-sync": "^1.4.10"
+  }
+}
